sdk extraction > cleanup, fix packages
