+++ src/main/java/com/hivemq/edge/modules/adapters/PollingProtocolAdapter.java
@@ -1,7 +1,6 @@
 package com.hivemq.edge.modules.adapters;
 
 import com.hivemq.edge.modules.adapters.data.ProtocolAdapterDataSample;
-import com.hivemq.edge.modules.api.adapters.ProtocolAdapter;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
