+++ src/main/java/com/hivemq/edge/modules/adapters/ProtocolAdapter.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.edge.modules.api.adapters;
+package com.hivemq.edge.modules.adapters;
 
-import com.hivemq.edge.modules.adapters.model.ProtocolAdapterDiscoveryInput;
-import com.hivemq.edge.modules.adapters.model.ProtocolAdapterDiscoveryOutput;
+import com.hivemq.edge.modules.adapters.discovery.ProtocolAdapterDiscoveryInput;
+import com.hivemq.edge.modules.adapters.discovery.ProtocolAdapterDiscoveryOutput;
 import com.hivemq.edge.modules.adapters.model.ProtocolAdapterStartInput;
 import com.hivemq.edge.modules.adapters.model.ProtocolAdapterStartOutput;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
