+++ src/main/java/com/hivemq/edge/modules/adapters/ProtocolAdapterTag.java
@@ -1,0 +1,13 @@
+package com.hivemq.edge.modules.adapters;
+
+public enum ProtocolAdapterTag {
+    UDP,
+    TCP,
+    SERIAL,
+    INTERNET,
+    WEB,
+    IOT,
+    IIOT,
+    AUTOMATION,
+    FACTORY
+}
