+++ src/main/java/com/hivemq/edge/modules/adapters/data/ProtocolAdapterDataSample.java
@@ -2,7 +2,7 @@ package com.hivemq.edge.modules.adapters.data;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.hivemq.edge.modules.config.AdapterSubscription;
+import com.hivemq.edge.modules.adapters.config.AdapterSubscription;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 
 import java.util.List;
