+++ src/main/java/com/hivemq/edge/modules/adapters/discovery/ProtocolAdapterDiscoveryOutput.java
@@ -13,28 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.edge.modules.config;
+package com.hivemq.edge.modules.adapters.discovery;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 
-/** Marker interface **/
 
-public interface ProtocolAdapterConfig {
+public interface ProtocolAdapterDiscoveryOutput {
 
-    String ID_REGEX = "^([a-zA-Z_0-9-_])*$";
+    @NotNull NodeTree getNodeTree();
 
-    int PORT_MIN = 1;
-    int PORT_MAX = 65535;
-    int DEFAULT_POLLING_INTERVAL = 1000;
-    int DEFAULT_MAX_POLLING_ERROR_BEFORE_REMOVAL = 10;
-
-    @NotNull String getId();
-
-     default int getPollingIntervalMillis() {
-       return 1000;
-    }
-
-     default int getMaxPollingErrorsBeforeRemoval() {
-        return -1;
-    }
 }
