+++ src/main/java/com/hivemq/edge/modules/adapters/factories/AdapterSubscriptionFactory.java
@@ -1,8 +1,8 @@
 package com.hivemq.edge.modules.adapters.factories;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.hivemq.edge.modules.config.AdapterSubscription;
-import com.hivemq.edge.modules.config.UserProperty;
+import com.hivemq.edge.modules.adapters.config.AdapterSubscription;
+import com.hivemq.edge.modules.adapters.config.UserProperty;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.annotations.Nullable;
