+++ src/main/java/com/hivemq/edge/modules/adapters/factories/EventBuilderFactory.java
@@ -1,6 +1,6 @@
 package com.hivemq.edge.modules.adapters.factories;
 
-import com.hivemq.edge.modules.api.events.model.EventBuilder;
+import com.hivemq.edge.modules.events.model.EventBuilder;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 
 public interface EventBuilderFactory {
