+++ src/main/java/com/hivemq/edge/modules/adapters/model/ProtocolAdapterInput.java
@@ -15,11 +15,11 @@
  */
 package com.hivemq.edge.modules.adapters.model;
 
+import com.hivemq.edge.modules.adapters.config.ProtocolAdapterConfig;
 import com.hivemq.edge.modules.adapters.factories.AdapterFactories;
-import com.hivemq.edge.modules.adapters.metrics.ProtocolAdapterMetricsHelper;
-import com.hivemq.edge.modules.api.adapters.ModuleServices;
-import com.hivemq.edge.modules.api.adapters.ProtocolAdapterState;
-import com.hivemq.edge.modules.config.ProtocolAdapterConfig;
+import com.hivemq.edge.modules.adapters.metrics.ProtocolAdapterMetricsService;
+import com.hivemq.edge.modules.adapters.services.ModuleServices;
+import com.hivemq.edge.modules.adapters.state.ProtocolAdapterState;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 
 public interface ProtocolAdapterInput<E extends ProtocolAdapterConfig> {
@@ -33,5 +33,5 @@ public interface ProtocolAdapterInput<E extends ProtocolAdapterConfig> {
 
     @NotNull AdapterFactories adapterFactories();
 
-    @NotNull ProtocolAdapterMetricsHelper getProtocolAdapterMetricsHelper();
+    @NotNull ProtocolAdapterMetricsService getProtocolAdapterMetricsHelper();
 }
