+++ src/main/java/com/hivemq/edge/modules/adapters/model/ProtocolAdapterStartInput.java
@@ -15,7 +15,7 @@
  */
 package com.hivemq.edge.modules.adapters.model;
 
-import com.hivemq.edge.modules.api.adapters.ModuleServices;
+import com.hivemq.edge.modules.adapters.services.ModuleServices;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 
 public interface ProtocolAdapterStartInput {
