+++ src/main/java/com/hivemq/edge/modules/adapters/services/ProtocolAdapterPublishService.java
@@ -13,21 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.edge.modules.adapters.model;
+package com.hivemq.edge.modules.adapters.services;
 
+import com.hivemq.edge.modules.adapters.ProtocolAdapterPublishBuilder;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
-import com.hivemq.extension.sdk.api.annotations.Nullable;
 
-public interface NodeTree {
+public interface ProtocolAdapterPublishService {
 
-    void addNode(
-            @NotNull String id,
-            @NotNull String name,
-            @NotNull String description,
-            @Nullable String parentId,
-            @NotNull NodeType nodeType,
-            boolean selectable);
-
+    @NotNull ProtocolAdapterPublishBuilder publish();
 }
-
-
