+++ src/main/java/com/hivemq/edge/modules/events/model/Event.java
@@ -1,14 +1,13 @@
-package com.hivemq.edge.modules.api.events.model;
+package com.hivemq.edge.modules.events.model;
 
-import com.hivemq.edge.model.Identifiable;
-import com.hivemq.edge.model.TypeIdentifier;
+import com.hivemq.edge.modules.events.EventService;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.annotations.Nullable;
 
 /**
- * Interface for events for the {@link com.hivemq.edge.modules.api.events.EventService}.
+ * Interface for events for the {@link EventService}.
  */
-public interface Event extends Identifiable {
+public interface Event {
     @NotNull SEVERITY getSeverity();
 
     @NotNull String getMessage();
@@ -23,8 +22,11 @@ public interface Event extends Identifiable {
 
     @Nullable TypeIdentifier getSource();
 
-    @Override
-    TypeIdentifier getIdentifier();
+    /**
+     * Represents a uniquely identifiable object in the system.
+     * @return The system-wide identifier of the object
+     */
+    @NotNull TypeIdentifier getIdentifier();
 
     public enum SEVERITY {
         INFO,
